+++ CHANGELOG.md
@@ -7,6 +7,10 @@ Versioning].
 
 ## [Unreleased]
 
+- _No changes yet_
+
+## [0.3.2] - 2020-03-13
+
 - Re-enable the Action from a Pull Request comment. ([#129])
 - Fix default commit message description. ([#130])